% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.R
\name{arc}
\alias{arc}
\title{Active learning using regularized cox models}
\usage{
arc(filename, regtype, metric, nfolds, alpha, lambda)
}
\arguments{
\item{filename}{Input matrix. Each row is an observation vector.}

\item{regtype}{Type of regularization to apply {"Lasso", "Enet", "Lapnet", "Kernelnet"}}

\item{metric}{Type of evaluation metric to fetch {"AUC", "MSE"}}

\item{nfolds}{Number of folds for cross validation}

\item{alpha}{ratio between L1 and Laplacian for "Lapnet", or between L1 and L2 for "Enet". Not needed for "Lasso". Value should be between 0 and 1}

\item{lambda}{a user supplied decreasing sequence. can range from 0.01 to 1}
}
\description{
Active learning using regularized cox models
}

