% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox-Regcox.R
\name{regcox}
\alias{regcox}
\title{Learn a Cox model with regularization}
\usage{
regcox(x, y, t, regtype, nfolds, standardize = TRUE, stdbeta = TRUE,
  lambda1 = 0, lambda2 = 0, alpha = 0.5, lambda = 0, tol = 0.001,
  itermax, weight = 1)
}
\arguments{
\item{x}{Input matrix. Each row is an observation vector.}

\item{y}{Response variable. y should be a two-column matrix with columns named 'time' and 'status'. The latter is a binary variable, with '1' indicating event, and '0' indicating right censored}

\item{t}{Time variable used for risk set calculation}

\item{regtype}{Type of regularization to apply {"Lasso", "Enet", "Lapnet", "Kernelnet", "Oscar", "Fear", "Cocktail"}}

\item{nfolds}{No of folds for cross validation}

\item{standardize}{Logical flag for standardizing x using z-score normalization. default = TRUE}

\item{stdbeta}{Logical flag for standardizing beta using min-max normalization. default=TRUE. Must be true for "Fear" and "Cocktail"}

\item{lambda1}{Additional parameter for Oscar regularization}

\item{lambda2}{Additional parameter for Oscar regularization}

\item{alpha}{ratio between L1 and Laplacian for "Lapnet". This value should be between 0 and 1}

\item{lambda}{Additional parameter for Fear regularization}

\item{tol}{Convergence tolerance}

\item{itermax}{Maximum number of iterations needed for convergence}

\item{weight}{Weight parameter for Cocktail regularization}
}
\description{
Learn a Cox model with regularization
}

